Change: Make callsign parser way more permissive
While it might be instructive to try to build a parser which matches
all varieties of acceptable callsigns, its going to be tremendously
involved, so instead, we'll go the other way, and start with something
that will match most things, and then add to it.  For our purposes it's
generally better to match more than less, and we can refine for now.

The currently failing test drives a complication - we need to handle
suffixes like /p /mm /qrp etc.